Expand ruff's rule set (B, C4, SIM, RUF, PERF) and fix what it found
The previous select=["E","F","I","UP"] missed exactly the categories
most likely to catch real bugs and efficiency antipatterns, as opposed
to pure style -- confirmed by what turned up once enabled:

- 18 stale `# noqa` comments suppressing rule codes (A002, N802, BLE001,
  ARG002) that were never in this project's select list to begin with --
  dead weight accumulating in the tree, auto-removed (RUF100).
- A genuine mutable-class-default: MediaCard._PLACEHOLDER_GLYPH was a
  plain dict class attribute -- annotated ClassVar[dict[str, str]] to
  make the "shared, never mutated" intent explicit (RUF012).
- Two zip() calls over lists that are only ever the same length by
  construction now say so explicitly (strict=True) -- a future length
  mismatch becomes a loud ValueError instead of a silently truncated
  loop (B905).
- try/except-in-loop in the personalized-shelves validation loop and the
  gamepad-device probe loop, both extracted into a per-item/per-path
  helper function so the try/except no longer sits lexically inside the
  loop (PERF203) -- real overhead on pre-3.11 Python, though this
  project's own tests run on 3.12.
- Three try/except/pass blocks (mpv's bounded wait, and CoverCache's
  best-effort file deletions during clear()/eviction) replaced with
  contextlib.suppress(...), which also clears two of the PERF203 hits
  above as a side effect since it's no longer a lexical try inside the
  loop.
- A handful of harmless leftovers: two unnecessary int(round(...)) casts
  (round() already returns int), one combinable nested async-with in
  updater.py, two en-dashes in test comments ruff flagged as ambiguous
  vs. a plain hyphen, and 14 unpacked-but-unused loop/tuple variables
  (mostly test_gamepad.py's `gl, actions = listener` where `actions`
  was never read) prefixed with `_`.

+++ tests/test_ui/test_widgets.py
@@ -376,7 +376,7 @@ def test_media_card_no_graphics_effect_ever_installed(qtbot):
 
 def _expected_scrim_alpha():
     from sixpack.ui import theme
-    return int(round(255 * (1.0 - theme.UNFOCUSED_OPACITY)))
+    return round(255 * (1.0 - theme.UNFOCUSED_OPACITY))
 
 
 def test_media_card_no_graphics_effect_used_for_dim(qtbot):
